[vision] Add base of Jetson vision app
Includes basic web api for setting camera
settings (exposure, thresholding?, etc)
as well as start of interface to c290 camera.

This will probably be done by interfacing to
v4l2-ctl but may use the v4l2 python bindings.
Not sure yet.

This commit also includes the single file
source of bottle, a small python web
framework. I included this in here so we don't have to
pip install anything on the Jetson.